Onstar

+++ selfdrive/ui/qt/onroad.h
@@ -171,9 +171,9 @@ private:
   ExperimentalButton *experimental_btn;
   QPixmap dm_img;
   float speed;
-  int onstar_gps_longitude;
-  int onstar_gps_latitude;
-  int onstar_gps_altitude;
+  float onstar_gps_longitude;
+  float onstar_gps_latitude;
+  float onstar_gps_altitude;
   QString speedUnit;
   float setSpeed;
   float speedLimit;
